update blue_current integration
