+++ homeassistant/components/blue_current/const.py
@@ -2,10 +2,9 @@
 
 import logging
 
-from homeassistant.const import Platform
-
 DOMAIN = "blue_current"
 
 LOGGER = logging.getLogger(__package__)
 
-PLATFORMS = [Platform.SENSOR]
+EVSE_ID = "evse_id"
+MODEL_TYPE = "model_type"
