+++ homeassistant/components/blue_current/translations/en.json
@@ -1,46 +1,46 @@
 {
-  "config": {
-      "abort": {
-          "already_configured": "Device is already configured"
-      },
-      "error": {
-          "already_connected": "IP is already connected",
-          "cannot_connect": "Failed to connect",
-          "invalid_token": "Invalid token",
-          "limit_reached": "Request limit reached",
-          "no_cards_found": "No charge cards found",
-          "unknown": "Unexpected error"
-      },
-      "step": {
-          "user": {
-              "data": {
-                  "api_token": "API Token"
-              },
-              "description": "Enter your Blue Current api token",
-              "title": "Authentication"
-          }
-      }
-  },
-  "entity": {
-    "sensor": {
-      "activity": {
-        "state": {
-          "available": "Available",
-          "charging": "Charging",
-          "unavailable": "Unavailable",
-          "error": "Error",
-          "offline": "Offline"
+    "config": {
+        "abort": {
+            "already_configured": "Device is already configured"
+        },
+        "error": {
+            "already_connected": "IP is already connected",
+            "cannot_connect": "Failed to connect",
+            "invalid_token": "Invalid token",
+            "limit_reached": "Request limit reached",
+            "no_cards_found": "No charge cards found",
+            "unknown": "Unexpected error"
+        },
+        "step": {
+            "user": {
+                "data": {
+                    "api_token": "API Token"
+                },
+                "description": "Enter your Blue Current api token",
+                "title": "Authentication"
+            }
         }
-      },
-      "vehicle_status": {
-        "state": {
-          "standby": "Standby",
-          "vehicle_detected": "Detected",
-          "ready": "Ready",
-          "no_power": "No power",
-          "vehicle_error": "Error"
+    },
+    "entity": {
+        "sensor": {
+            "activity": {
+                "state": {
+                    "available": "Available",
+                    "charging": "Charging",
+                    "error": "Error",
+                    "offline": "Offline",
+                    "unavailable": "Unavailable"
+                }
+            },
+            "vehicle_status": {
+                "state": {
+                    "no_power": "No power",
+                    "ready": "Ready",
+                    "standby": "Standby",
+                    "vehicle_detected": "Detected",
+                    "vehicle_error": "Error"
+                }
+            }
         }
-      }
     }
-  }
 }
