+++ requirements_test_all.txt
@@ -469,7 +469,7 @@ blebox-uniapi==2.2.0
 blinkpy==0.22.4
 
 # homeassistant.components.blue_current
-bluecurrent-api==1.0.1
+bluecurrent-api==1.0.3
 
 # homeassistant.components.bluemaestro
 bluemaestro-ble==0.2.3
